pyproject.toml
[project]
name = "home-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings",
    "pytest>=9.0.2",
    "aiosqlite",
    "pytest-asyncio",
    "pydantic-ai>=1.62.0",
    "python-telegram-bot>=21.6",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "import-linter>=2.10",
    "ty>=0.0.17",
    "pyyaml>=6.0",
]

[tool.ty.rules]
# pydantic-settings loads required fields from env vars at runtime;
# ty can't know this so we suppress the false positive
missing-argument = "ignore"

[tool.importlinter]
root_packages = ["home_agent"]

# Layer 0 — config: no internal imports allowed
[[tool.importlinter.contracts]]
name = "config is independent"
type = "forbidden"
source_modules = ["home_agent.config"]
forbidden_modules = [
    "home_agent.db",
    "home_agent.profile",
    "home_agent.history",
    "home_agent.agent",
    "home_agent.bot",
    "home_agent.tools",
    "home_agent.mcp",
]

# Layer 1 — db: only imports config and shared
[[tool.importlinter.contracts]]
name = "db only imports config and shared"
type = "forbidden"
source_modules = ["home_agent.db"]
forbidden_modules = [
    "home_agent.profile",
    "home_agent.history",
    "home_agent.agent",
    "home_agent.bot",
    "home_agent.tools",
    "home_agent.mcp",
]

# Layer 1b — history: only imports config, db, and shared
[[tool.importlinter.contracts]]
name = "history only imports config and db"
type = "forbidden"
source_modules = ["home_agent.history"]
forbidden_modules = [
    "home_agent.agent",
    "home_agent.bot",
    "home_agent.tools",
    "home_agent.mcp",
]

# Layer 2 — mcp: no imports from agent, bot, main
# NOTE: When home_agent.tools is created (step 1.9+), replace the active contract
# below with this combined one that covers both tools and mcp together:
# [[tool.importlinter.contracts]]
# name = "tools and mcp do not import agent or bot"
# type = "forbidden"
# source_modules = ["home_agent.tools", "home_agent.mcp"]
# forbidden_modules = [
#     "home_agent.agent",
#     "home_agent.bot",
#     "home_agent.main",
# ]

[[tool.importlinter.contracts]]
name = "mcp does not import agent or bot"
type = "forbidden"
source_modules = ["home_agent.mcp"]
forbidden_modules = [
    "home_agent.agent",
    "home_agent.bot",
    "home_agent.main",
]

# Layer 3 — agent and bot: no circular imports with each other or main
[[tool.importlinter.contracts]]
name = "agent does not import bot or main"
type = "forbidden"
source_modules = ["home_agent.agent"]
forbidden_modules = [
    "home_agent.bot",
    "home_agent.main",
]

[[tool.importlinter.contracts]]
name = "bot does not import main"
type = "forbidden"
source_modules = ["home_agent.bot"]
forbidden_modules = ["home_agent.main"]

# Layer 4 — main: composition root; no other module should import it
[[tool.importlinter.contracts]]
name = "other modules do not import main"
type = "forbidden"
source_modules = [
    "home_agent.config",
    "home_agent.db",
    "home_agent.profile",
    "home_agent.history",
    "home_agent.agent",
    "home_agent.bot",
    "home_agent.mcp",
]
forbidden_modules = ["home_agent.main"]

